删除流程时，判断节点表是否存在

+++ jflow-core/src/main/java/BP/WF/WorkFlow.java
@@ -473,6 +473,9 @@ public class WorkFlow
 		{
 			try
 			{
+				 if (DBAccess.IsExitsObject("ND" + nd.getNodeID()) == false)
+                     continue;
+				 
 				DBAccess.RunSQL("DELETE FROM ND" + nd.getNodeID() + " WHERE OID=" + workid + " OR FID=" + workid);
 			}
 			catch (RuntimeException ex)
